+++ scripts/action-forge.js
@@ -24,7 +24,7 @@ Hooks.once("init", () => {
     });
   }
 
-  console.info(`PF2E Action Forge | Initialized ${module?.version ?? "0.1.0-dev.9"}`);
+  console.info(`PF2E Action Forge | Initialized ${module?.version ?? "0.1.0-dev.10"}`);
 });
 
 Hooks.on("getSceneControlButtons", (controls) => {